feat: file scanning (#5)
* feat: file scanning

This adds file scanning and initial matrix creation.

Begins #4

* feat: matrix creation

* feat: begins plugin communication with core

This is a semi-working commit to create a place in-time for our plugin operations code before a refactor. The intention is to put this on the history to then go back and make the plugin to core communication happen in a more file direct way.

* refactor: starting to convert to file based communication