refactor: changing how the plugins communicate with core (#7)
* refactor: changing how the plugins communicate with core

This will finish the refactor of using the cache to write JSON files for communication between the core and plugin system. This is more reliable than buffer based communication.

Resolves #4

* feat: adding bundled rust scanner

This finished the conversion of buffer communication to file-based and additionally adds the rust scanner. Now this project is capable of scanning itself.